Safari-611.1.10/LayoutTests/fast/events/popup-blocked-from-untrusted-click-event-on-anchor-expected.txt

test
PASS win is null
